backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
django-cors-headers==3.13.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
filetype==1.2.0
flake8==7.1.1
gunicorn==20.1.0
isort==5.10.1
packaging==24.1
Pillow==10.4.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
sqids==0.5.0
webcolors==1.11.1
